+++ packages/keyring-api/src/utils/index.ts
@@ -1,2 +1,3 @@
+export * from './pagination';
 export * from './types';
 export * from './typing';
